requirements.txt
# datasets
# requests
groq
pandas
python_dotenv
replicate
tavily-python

